requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0
bcrypt==4.1.2
itsdangerous==2.1.2
python-dotenv==1.0.0
requests==2.31.0
